Make refinedstorage2-resource-api succeed Pitest requirements
