+++ refinedstorage2-resource-api/src/main/java/com/refinedmods/refinedstorage2/api/resource/list/ResourceListImpl.java
@@ -24,8 +24,6 @@ public class ResourceListImpl<T> implements ResourceList<T> {
 
     @Override
     public ResourceListOperationResult<T> add(final T resource, final long amount) {
-        ResourceAmount.validate(resource, amount);
-
         final ResourceAmount<T> existing = entries.get(resource);
         if (existing != null) {
             return addToExisting(existing, amount);
