+++ refinedstorage2-resource-api/src/test/java/com/refinedmods/refinedstorage2/api/resource/list/ProxyResourceListTest.java
@@ -1,0 +1,9 @@
+package com.refinedmods.refinedstorage2.api.resource.list;
+
+class ProxyResourceListTest extends AbstractResourceListTest {
+    @Override
+    protected ResourceList<String> createList() {
+        return new AbstractProxyResourceList<>(new ResourceListImpl<>()) {
+        };
+    }
+}
